Refactored code to use ResponseEntity as a return type in the controllers. Centralized exception throwing to service classes and handling in the controllers. Removed any exception handling logic from the UI layer.

+++ src/main/java/controller/OrganizerController.java
@@ -3,6 +3,7 @@ package controller;
 import dto.ConferenceDTO;
 import dto.ConferenceFormDTO;
 import exception.*;
+import response.ResponseEntity;
 import service.ConferenceService;
 import service.UserService;
 import util.LoggerUtil;
@@ -18,55 +19,64 @@ public class OrganizerController {
         this.conferenceService = conferenceService;
     }
 
-    public Optional<ConferenceDTO> getManagedConference(String conferenceId) {
+    public ResponseEntity<ConferenceDTO> getManagedConference(String conferenceId) {
         try {
             ConferenceDTO conferenceDTO = conferenceService.findById(conferenceId);
-            return Optional.of(conferenceDTO);
+            return ResponseEntity.success(conferenceDTO);
         } catch (ConferenceNotFoundException e) {
             LoggerUtil.getInstance().logError("Conference with id '" + conferenceId + "' could not be found.");
-            return Optional.empty();
+            return ResponseEntity.error(e.getMessage());
         }
     }
 
-    public List<ConferenceDTO> getManagedConferences(String email) {
+    public ResponseEntity<List<ConferenceDTO>> getManagedConferences(String email) {
         try {
             Set<String> conferenceIds = userService.findManagedConferencesForOrganizer(email);
-            return conferenceService.findByIds(conferenceIds);
-
+            List<ConferenceDTO> conferenceDTOS = conferenceService.findByIds(conferenceIds);
+            return ResponseEntity.success(conferenceDTOS);
         } catch (InvalidUserRoleException | UserNotFoundException e) {
             LoggerUtil.getInstance().logError("Failed to retrieve managed conferences for user with email '" + email + "' due to the following reason: " + e.getMessage());
-            return Collections.emptyList();
+            return ResponseEntity.error("Error retrieving managed conferences. Please try again later.");
         }
     }
 
-    public void validateConferenceData(ConferenceFormDTO conferenceFormDTO) {
+    public ResponseEntity<Void> validateConferenceData(ConferenceFormDTO conferenceFormDTO) {
         // implement validation logic
         String name = conferenceFormDTO.getName();
         Date startDate = conferenceFormDTO.getStartDate(), endDate = conferenceFormDTO.getEndDate();
 
         if (conferenceService.isNameTaken(name)) {
             LoggerUtil.getInstance().logWarning("Validation failed for conference creation. Conference name '" + name + "' is already taken.");
-            throw ConferenceCreationException.nameTaken();
+            return ResponseEntity.error("The provided conference name is already taken.");
         }
 
         if (!conferenceService.isTimePeriodAvailable(startDate, endDate)) {
             LoggerUtil.getInstance().logWarning("Validation failed for conference creation. Dates provided for the conference are not available..");
-            throw ConferenceCreationException.dateUnavailable();
+            return ResponseEntity.error("Another conference is already registered to take place within the time period provided. Please choose different dates.");
         }
 
         LoggerUtil.getInstance().logInfo("Validation successful for conference: " + name + " with dates: " + startDate + " - " + endDate);
+        return ResponseEntity.success();
     }
 
-    public void createConference(ConferenceFormDTO conferenceFormDTO) {
-        String conferenceId = conferenceService.create(conferenceFormDTO);
+    public ResponseEntity<Void> createConference(ConferenceFormDTO conferenceFormDTO) {
+        String conferenceId = null;
         try {
+            conferenceId = conferenceService.create(conferenceFormDTO);
             userService.addNewManagedConferenceForOrganizer(conferenceFormDTO.getOrganizerId(), conferenceId);
-        } catch (Exception e) {
-            // rolling back conference creation if adding the conference to the organizer's managed conferences failed
-            conferenceService.deleteById(conferenceId);
 
-            // propagate saving data exception upwards for user friendly error
-            if (e instanceof SavingDataException) throw e;
+            return ResponseEntity.success();
+        } catch (SavingDataException e) {
+            // rolling back conference creation if adding the conference to the organizer's managed conferences failed
+            if (conferenceId != null) {
+                conferenceService.deleteById(conferenceId);
+            }
+
+            LoggerUtil.getInstance().logError("Error occurred during conference creation or adding conference to organizer's managed conferences: " + e.getMessage());
+            return ResponseEntity.error("An error occurred while creating the conference. Please try again later.");
+        } catch (Exception e) {
+            LoggerUtil.getInstance().logError("An unexpected error occurred during conference creation: " + e.getMessage());
+            return ResponseEntity.error("An unexpected error occurred. Please try again later.");
         }
     }
 }
